backend/requirements.txt
fastapi
uvicorn
python-multipart
qrcode
pillow
